options for 1-5 custom targets in the nav ring
When user chooses 4 or 5 targets the ring layout is raised vertically
and the ring diameter is shrunk a bit to accommodate the targets at the
9 and 3.

I've also got layouts in for Small tabs (grouper) .
I need a HDPI device and a full tab (if we have one) to test these and
let me know if anything needs fixing.

So since there is no git history for this fiasco let me say thanks to
Kevin Liu and CNA for the running head start on this.